+++ brink.sh
@@ -129,6 +129,10 @@ clean_build() {
     delete_folder ${DIST_FOLDER}
     echo "Removing publish..."
     delete_folder 'publish'
+    echo "Removing node_modules..."
+    delete_folder node_modules
+    echo "Removing web build"
+    delete_folder chevah/server/static/build/
 
     # In some case pip hangs with a build folder in temp and
     # will not continue until it is manually removed.
@@ -247,7 +251,7 @@ update_path_variables() {
     export CHEVAH_OS=${OS}
     export CHEVAH_ARCH=${ARCH}
     export CHEVAH_CACHE=${CACHE_FOLDER}
-    export PIP_INDEX=${PIP_INDEX}
+    export PIP_INDEX_URL=${PIP_INDEX_URL}
 
 }
 
@@ -301,9 +305,7 @@ pip_install() {
     set +e
     ${PYTHON_BIN} -m \
         pip install \
-            --trusted-host bin.chevah.com \
-            --trusted-host deag.chevah.com \
-            --index-url=$PIP_INDEX \
+            --index-url=$PIP_INDEX_URL \
             $1
 
     exit_code=$?
@@ -716,6 +718,9 @@ detect_os() {
                         if [ ${os_version_chevah%%04} == ${os_version_chevah} \
                             -o $(( ${os_version_chevah:0:2} % 2 )) -ne 0 ]; then
                             check_linux_glibc
+                        elif [ ${os_version_chevah} == "2204" ]; then
+                            # OpenSSL 3.0.x not supported by cryptography 3.3.x.
+                            check_linux_glibc
                         fi
                         set_os_if_not_generic "ubuntu" $os_version_chevah
                         ;;
@@ -786,17 +791,6 @@ detect_os() {
             ;;
         "amd64"|"x86_64")
             ARCH="x64"
-            case "$OS" in
-                win)
-                    # 32bit build on Windows 2016, 64bit otherwise.
-                    # Should work with a l10n pack too (tested with French).
-                    win_ver=$(systeminfo.exe | head -n 3 | tail -n 1 \
-                        | cut -d ":" -f 2)
-                    if [[ "$win_ver" =~ "Microsoft Windows Server 2016" ]]; then
-                        ARCH="x86"
-                    fi
-                    ;;
-            esac
             ;;
         "aarch64")
             ARCH="arm64"